requirements.txt
pyyaml==6.0.2
polars==1.9.0